requirements.txt
langchain
langchain-openai
faiss-cpu
pymupdf
streamlit
boto3
python-dotenv